source_code/requirements.txt
librosa
numpy 
keras.models 
flask 
flask_restful
librosa
pandas 
joblib
matplotlib
tensorflow